chore(bot.py): ignore false-positive Pylint warning
This line was resulting in: `W0221: Number of parameters was 0 in 'BotBase.on_command_error' and is now 3 in overriding 'GitCordBot.on_command_error' method (arguments-differ)`

We're overriding `on_command_error` exactly as described by the d.py documentation and source code, so I'm not sure why Pylint thinks `BotBase.on_command_error` has 0 parameters. Considering this a false positive and disabling it for the line.